requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
gunicorn>=21.0.0
whitenoise>=6.0.0
python-decouple>=3.8
reportlab==4.0.4
openpyxl==3.1.2
python-docx==1.1.0